Use float for letterSpacing in font rendering
Change letterSpacing parameter from int to float for font rendering APIs to allow sub-pixel/ finer spacing control. Updated signatures in FontResourceGenerator.generateNativeImage and TextureList.renderFont, and minor comment cleanup in TextureList. Note: call sites will need to be updated to pass a float value; runtime behavior should improve text layout precision.

+++ fabric/src/main/java/top/xfunny/mod/client/FontResourceGenerator.java
@@ -24,7 +24,7 @@ public class FontResourceGenerator implements IGui {
      * @param padding        文字内边距
      * @param letterSpacing  字符间距
      */
-    public static NativeImage generateNativeImage(String text, int textColor, Font font, float fontSize, int padding, int letterSpacing) {// 使用AI优化
+    public static NativeImage generateNativeImage(String text, int textColor, Font font, float fontSize, int padding, float letterSpacing) {// 使用AI优化
         // 基础缩放值
         int baseScale = (int) Math.pow(2, Config.getClient().getDynamicTextureResolution() + 5);
